Move test server setup to own file
